Fix unintended behavior
In previous versions, if a user pressed a numeric after a result was
displayed, and hit the = key, the calculator would use the previous
operator to determine th next result. For example: 55 + 45 = 100, user
then enters 80 without an operator resulting in 180 being displayed.
Now, the calculator will clear all data in this scenario.